moves components to separate folder
