+++ app/components/Home.js
@@ -1,0 +1,7 @@
+import React, { Component } from 'react'
+
+const Home = () => (
+	<div id="home-container">Welcome to my personal page!  As you can see, this is very much a work in progress.  I invite you to click on the 'Resume' tab to learn a little about my work.</div>
+)
+
+export default Home
